Fixed the crashing issues with the joints
Apparently making a loop from the lower arm back to the torso using a free joint fixed the crashing. That free joint had to be added to the lower arm as an ImpulseJoint though. I'm not sure how long this solution is gonna last, but at least I get to enjoy the flawlessness of multibody joints for a while.